update getTransferRemoteWithKernel tests
